refactor: organize src/components into logical subdirectories
