+++ src/app/(home)/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 import App from "@/components/App";
-import { AppErrorBoundary } from "@/components/ErrorBoundary";
+import { AppErrorBoundary } from "@/components/shared/ErrorBoundary";
 import { parseMetadataParams } from "@/lib/metadata";
 
 interface Props {
