+++ src/app/guides/how-interest-works/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
-import { AppErrorBoundary } from "@/components/ErrorBoundary";
 import { InterestGuide } from "@/components/guides/how-interest-works/InterestGuide";
+import { AppErrorBoundary } from "@/components/shared/ErrorBoundary";
 
 export const metadata: Metadata = {
   alternates: { canonical: "/guides/how-interest-works" },
