+++ src/app/guides/rpi-vs-cpi/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
-import { AppErrorBoundary } from "@/components/ErrorBoundary";
 import { RpiVsCpiGuide } from "@/components/guides/rpi-vs-cpi/RpiVsCpiGuide";
+import { AppErrorBoundary } from "@/components/shared/ErrorBoundary";
 
 export const metadata: Metadata = {
   alternates: { canonical: "/guides/rpi-vs-cpi" },
