--- src/components/DateInput.tsx
@@ -1,60 +1,0 @@
-"use client";
-
-import { Calendar03Icon } from "@hugeicons/core-free-icons";
-import { HugeiconsIcon } from "@hugeicons/react";
-import type { DateInputProps } from "@/types/input";
-import { Button } from "@/components/ui/button";
-import { Calendar } from "@/components/ui/calendar";
-import { Label } from "@/components/ui/label";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
-import { formatDateLong } from "@/lib/date-utils";
-import { cn } from "@/lib/utils";
-
-export function DateInput({
-  id,
-  label,
-  helperText,
-  value,
-  onChange,
-}: DateInputProps) {
-  return (
-    <div className="space-y-2">
-      <Label htmlFor={id}>{label}</Label>
-      <Popover>
-        <PopoverTrigger
-          render={
-            <Button
-              id={id}
-              variant="outline"
-              className={cn(
-                "w-full justify-start text-left font-normal",
-                !value && "text-muted-foreground",
-              )}
-            >
-              <HugeiconsIcon icon={Calendar03Icon} className="mr-2 size-4" />
-              {value ? formatDateLong(value) : "Pick a date"}
-            </Button>
-          }
-        />
-        <PopoverContent className="w-auto p-0" align="start">
-          <Calendar
-            mode="single"
-            selected={value ?? undefined}
-            onSelect={(date) => {
-              onChange(date ?? null);
-            }}
-          />
-        </PopoverContent>
-      </Popover>
-      {helperText && (
-        <p className="text-sm text-muted-foreground">{helperText}</p>
-      )}
-    </div>
-  );
-}
-
-export default DateInput;
