+++ src/components/home/CurrencyInput.tsx
