+++ src/components/home/ResultSummary.tsx
