+++ src/components/overpay/OverpayPage.tsx
@@ -5,14 +5,14 @@ import { OverpayComparisonChart } from "./OverpayComparisonChart";
 import { OverpayPrimaryInputs } from "./OverpayPrimaryInputs";
 import { OverpaySummaryCards } from "./OverpaySummaryCards";
 import { OverpayVerdict } from "./OverpayVerdict";
-import type { InputMode } from "@/components/InputPanel";
+import type { InputMode } from "@/components/home/InputPanel";
 import type { Preset } from "@/lib/presets";
-import { AssumptionsCallout } from "@/components/AssumptionsCallout";
-import { Breadcrumb } from "@/components/Breadcrumb";
-import { Footer } from "@/components/Footer";
-import { Header } from "@/components/Header";
-import { InputPanel } from "@/components/InputPanel";
-import { PlanFromQuery } from "@/components/PlanFromQuery";
+import { InputPanel } from "@/components/home/InputPanel";
+import { Breadcrumb } from "@/components/layout/Breadcrumb";
+import { Footer } from "@/components/layout/Footer";
+import { Header } from "@/components/layout/Header";
+import { AssumptionsCallout } from "@/components/shared/AssumptionsCallout";
+import { PlanFromQuery } from "@/components/shared/PlanFromQuery";
 import { Skeleton } from "@/components/ui/skeleton";
 import { useLoanActions, useLoanConfigState } from "@/context/LoanContext";
 import { useOverpayAnalysis } from "@/hooks/useOverpayAnalysis";
