+++ src/components/shared/PlanFromQuery.tsx
