+++ src/components/shared/SalaryGrowthBadge.tsx
